Use stack view
This eliminates the need for you to set x and y constraints for the subviews.

You can decide whether you want it even distributed across the container view, or whether you just want fixed spacing. This happens to do the evenly distribution of the subviews.

It doesn't, though, gracefully handle if the number of subviews exceeds what can fit horizontally. For that, you'd use `UICollectionView`.