Use UICollectionView
In this case, because I gathered that you might have something else in the main view controller, I implemented the friend icon view controller as an "embedded" view controller (so the main view controller isn't polluted with a bunch of stuff related to managing the friend view controller list).

There's too much here for me to enumerate all the details to get the collection view controller to work properly, but if you're not familiar with collection views, Google "UICollectionView tutorial" or "UICollectionViewController tutorial" and you should find good links. Once you've gained a degree of confidence with how collection views work, you can come back to this and grok what's going on here.